packages/server/cloud_functions/registration_counts/requirements.txt
firebase_admin
google-cloud-monitoring